+++ app/Http/Controllers/TournamentController.php
@@ -5,9 +5,10 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 use App\Models\Tournament;
-use App\Models\Tournament_Avatar;
+use App\Models\tournament_avatar;
 use App\Models\Game;
 use App\Models\User;
+use App\Models\Points;
 use Illuminate\Support\Facades\Auth;
 
 
@@ -78,7 +79,11 @@ class TournamentController extends Controller
      */
     public function edit($id)
     {
-        //
+        $user = Auth::user();
+        $tournament_avatars = tournament_avatar::all();
+        $tournament = Tournament::findOrFail($id);
+        $points = Points::where('user_id',$user->id)->get();
+        return view('console.tournaments.edit',compact('tournament','tournament_avatars'),compact('points'));
     }
 
     /**
@@ -90,7 +95,23 @@ class TournamentController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        $tournament = Tournament::findOrFail($id);
+        $formFields=$request->validate([
+            'name'=>'required',
+            'type'=>'required',
+            'fees'=>'required',
+            'closing_time'=>'required',
+            'team_size'=>'required',
+            'prize_pool'=>'required',
+            'first_prize'=>'required',
+            'second_prize'=>'required',
+            'third_prize'=>'required',
+            'rules'=>'required',
+            'image_id'=>'required',
+        ]);
+        $tournament->update($formFields + ['user_id'=>auth()->user()->id]);
+        return redirect('/dashboard')->with('message','Tournament has been updated successfully');
+
     }
 
     /**
@@ -102,8 +123,13 @@ class TournamentController extends Controller
     public function destroy($id)
     {
         $tournament = Tournament::findOrFail($id);
+        // Make sure logged in user is owner
+        if($tournament->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $tournament->delete();
         return redirect('/dashboard')->with('message','Tournament deleted succesfully');
-
     }
 }
+
+?>
